requirements-dev.txt
fastapi==0.109.2
uvicorn==0.27.1
python-sane==2.9.1
Pillow==9.5.0
python-lsp-server[all]
pytest==7.2.0
pytest-cov==2.11.1
pytest-mock==3.10.0
debugpy==1.6.7
